Moved logger to log-utils
