+++ src/language-js/sortTSPropertySignatures/index.ts
@@ -4,7 +4,7 @@ import {
   getContextGroups,
   getSpreadGroups,
   reorderValues
-} from "../common/sort-utils";
+} from "../utilities/sort-utils";
 
 export interface SortTSPropertySignaturesOptions {
   groups: ("null" | "undefined" | "*" | "function" | "object")[];
