+++ src/utilities/string-utils.ts
